[LIVY-572] Avoid usage of spark classes in ColumnBuffer
## What changes were proposed in this pull request?

The `ColumnBuffers` can be created both inside spark jobs and in the Livy server. The latter case happens when operation logs are returned and  fails before this patch because we are using Spark classes in this code after the refactor in LIVY-503. Unfortunately, we do not have test coverage for operation logs retrieval and this is the reason why this wasn't spot out earlier.

Since operation logs are retrieved by beeline for each query, this means that every query run through beeline fails, unless `livy.server.thrift.logging.operation.enabled` is set to `false`.

## How was this patch tested?

manual tests using beeline

Author: Marco Gaido <mgaido@apache.org>

Closes #162 from mgaido91/LIVY-572.
